A probar lo de crisitan

+++ src/java/Modelo/PrecioId.java
@@ -9,29 +9,29 @@ package Modelo;
 public class PrecioId  implements java.io.Serializable {
 
 
-     private short idestablecimiento;
-     private long idproducto;
+     private int idestablecimiento;
+     private int idproducto;
 
     public PrecioId() {
     }
 
-    public PrecioId(short idestablecimiento, long idproducto) {
+    public PrecioId(int idestablecimiento, int idproducto) {
        this.idestablecimiento = idestablecimiento;
        this.idproducto = idproducto;
     }
    
-    public short getIdestablecimiento() {
+    public int getIdestablecimiento() {
         return this.idestablecimiento;
     }
     
-    public void setIdestablecimiento(short idestablecimiento) {
+    public void setIdestablecimiento(int idestablecimiento) {
         this.idestablecimiento = idestablecimiento;
     }
-    public long getIdproducto() {
+    public int getIdproducto() {
         return this.idproducto;
     }
     
-    public void setIdproducto(long idproducto) {
+    public void setIdproducto(int idproducto) {
         this.idproducto = idproducto;
     }
 
@@ -41,14 +41,12 @@ public class PrecioId  implements java.io.Serializable {
 		 if ( (other == null ) ) return false;
 		 if ( !(other instanceof PrecioId) ) return false;
 		 PrecioId castOther = ( PrecioId ) other; 
-         
 		 return (this.getIdestablecimiento()==castOther.getIdestablecimiento())
  && (this.getIdproducto()==castOther.getIdproducto());
    }
    
    public int hashCode() {
          int result = 17;
-         
          result = 37 * result + this.getIdestablecimiento();
          result = 37 * result + (int) this.getIdproducto();
          return result;
